Corrected 5.4.1.4 so that INACTIVE is not added multiple times
Corrected 5.4.1.4 so that INACTIVE is not added multiple times. \s+ forced a
space following INACTIVE, and since the line we added did not (and probably
should not) have one, our line was re-added every time this was run.

Replaced with a regex that permits a space but does not force one, that
does force a following equals sign, and for good measure mathes until the
end of the line.