Make 5.4.2 check-mode compliant
It seems the ansible user module does not support locking accounts.
Instead of running usermod -L on all sytem accounts, I limit the
run to system accounts that are unlocked.

It is possible to do this without resorting to bash, but the way I
found was extremely inefficient (a task looping over each user on
the system, locked or not, with ssh and sftp etc.)