Make 3.3.1 and 3.3.2 check-mode compliant
Using sysctl_set will fail on unknown keys even if ignoreerrors is given
(this a bug in ansible, either doc or code). Correct this by reloading
instead of manipulating each value by itself. This is not perfect since
reload will not happen if the line is correct but sysctl has not been
reloaded. Contemplating logging an issue with ansible.